NL-2-sql-App/pyproject.toml
[project]
name = "nl2sql-app"
version = "0.1.0"
description = "Natural Language to SQL Query Processing with AI Agents"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.31.0",
    "chromadb>=0.4.22",
    "pydantic>=2.6.1",
    "sqlparse>=0.4.4",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.0",
    "pytest>=8.0.0",
    "llama-cpp-python>=0.2.39",
    "openai>=1.12.0",
    "tiktoken>=0.11.0"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"